NIFI-1762: Changed Java dependency to 1.8 instead of 1.7 and refactored StandardNiFiServiceFacade to make use of Lambda expressions to simplify code base. Also had to address a unit test because changing to Java 8 results in calls to assertEquals to become ambiguous. This closes #352

+++ nifi-nar-bundles/nifi-jms-bundle/nifi-jms-cf-service/src/test/java/org/apache/nifi/jms/cf/JMSConnectionFactoryProviderTest.java
@@ -69,9 +69,9 @@ public class JMSConnectionFactoryProviderTest {
         assertNotNull(cf);
         assertEquals("org.apache.nifi.jms.testcflib.TestConnectionFactory", cf.getClass().getName());
         assertEquals("myhost", this.get("getHost", cf));
-        assertEquals(1234, this.get("getPort", cf));
+        assertEquals(1234, ((Integer) this.get("getPort", cf)).intValue());
         assertEquals("foo", this.get("getFoo", cf));
-        assertEquals(3, this.get("getBar", cf));
+        assertEquals(3, ((Integer) this.get("getBar", cf)).intValue());
     }
 
     @Test(expected = AssertionError.class)
